比较不同分类器、特征；聚类版新闻分类；词性过滤

+++ segmenter.py
@@ -5,25 +5,40 @@ import os.path
 import MySQLdb
 import shutil
 import jieba
-from myutils import StopWord, StopSent, ArticleDB
+import jieba.posseg as pseg
+from myutils import StopWord, StopSent, ArticleDB, PosFilter
 
 
 class Segmenter:
-    def __init__(self, proj_name):
+    def __init__(self, proj_name, rm_exist=False):
         # 创建分词文件存放目录，已存在则删除重建
         self.proj_name = proj_name
         self.seg_dir = proj_name + "/seg"
         self.seg_join_dir = proj_name + "/seg_join"
-        shutil.rmtree(self.seg_dir, ignore_errors=True)
-        shutil.rmtree(self.seg_join_dir, ignore_errors=True)
-        os.makedirs(self.seg_dir)
-        os.makedirs(self.seg_join_dir)
+        self.tt_dir = proj_name + "/tt"  # title和tags的分词
+        self.pos_seg_dir = proj_name + "/pos_seg"
+        if rm_exist:
+            shutil.rmtree(self.seg_dir, ignore_errors=True)
+            shutil.rmtree(self.seg_join_dir, ignore_errors=True)
+            shutil.rmtree(self.tt_dir, ignore_errors=True)
+            shutil.rmtree(self.pos_seg_dir, ignore_errors=True)
+
+        try:
+            os.makedirs(self.seg_dir)
+            os.makedirs(self.seg_join_dir)
+            os.makedirs(self.tt_dir)
+            os.makedirs(self.pos_seg_dir)
+        except:
+            pass
 
         # 分句符号
         self.stop_sent = StopSent()
 
         # 停用词
         self.stop_word = StopWord("./stopwords_general.txt")
+
+        # 词性过滤
+        self.pos_filter = PosFilter()
 
         # 连接数据库，获取文章总数
         db = ArticleDB()
@@ -34,12 +49,10 @@ class Segmenter:
         except MySQLdb.Error, e:
             print "Mysql Error %d: %s" % (e.args[0], e.args[1])
 
-        # 分词工具初始化
-        # self.segtool = thulac.thulac("-seg_only -model_dir models/")
-        self.segtool = jieba
-
     # 分词
     def seg(self):
+        # self.segtool = thulac.thulac("-seg_only -model_dir models/")
+        segtool = pseg
         # 遍历所有文件，进行分词
         for i in range(1, self.doc_count + 1):
             print ("\r%d/%d" % (i, self.doc_count))
@@ -51,10 +64,53 @@ class Segmenter:
                 for line in txt_file.readlines():
                     sentences = self.seg_sentence(line)
                     for sentence in sentences:
-                        words = self.segtool.cut(sentence)
+                        words = segtool.cut(sentence)
                         words = [word.encode("utf-8") for word in words]  # jieba need, thulac skip this code
                         words = [word for word in words if not self.stop_word.is_stop_word(word)]
                         doc.append(" ".join(words) + "\n")
+                tmp_file.writelines(doc)
+            os.rename(tmp_name, seg_name)
+
+    # 分词(带词性过滤)
+    def pos_seg(self):
+        segtool = pseg
+        # 遍历所有文件，进行分词
+        for i in range(1, self.doc_count + 1):
+            print ("\r%d/%d" % (i, self.doc_count))
+            txt_name = "%s/txt/%d" % (self.proj_name, i)
+            tmp_name = "%s/pos_seg/%d.tmp" % (self.proj_name, i)
+            seg_name = "%s/pos_seg/%d" % (self.proj_name, i)
+            doc = []
+            with open(txt_name, "r") as txt_file, open(tmp_name, "w") as tmp_file:
+                for line in txt_file.readlines():
+                    sentences = self.seg_sentence(line)
+                    for sentence in sentences:
+                        words = segtool.cut(sentence)
+                        # jieba need the next line code, thulac skip this
+                        words = [(word.word.encode("utf-8"), word.flag.encode("utf-8")) for word in words]
+                        words = [word for word in words if not self.stop_word.is_stop_word(word[0])]
+                        words = [word for word in words if self.pos_filter.is_good_pos(word[1][0])]
+                        words = [word[0] for word in words]
+                        doc.append(" ".join(words) + "\n")
+                tmp_file.writelines(doc)
+            os.rename(tmp_name, seg_name)
+
+    # 对标题和标签分词
+    def seg_title_tags(self):
+        # 遍历所有文件，进行分词
+        for i in range(1, self.doc_count + 1):
+            print ("\r%d/%d" % (i, self.doc_count))
+            txt_name = "%s/txt/%d" % (self.proj_name, i)
+            tmp_name = "%s/tt/%d.tmp" % (self.proj_name, i)
+            seg_name = "%s/tt/%d" % (self.proj_name, i)
+            doc = []
+            with open(txt_name, "r") as txt_file, open(tmp_name, "w") as tmp_file:
+                lines = [txt_file.readline(), txt_file.readline()]
+                for sentence in lines:
+                    words = self.segtool.cut(sentence)
+                    words = [word.encode("utf-8") for word in words]  # jieba need, thulac skip this code
+                    words = [word for word in words if not self.stop_word.is_stop_word(word)]
+                    doc.append(" ".join(words) + "\n")
                 tmp_file.writelines(doc)
             os.rename(tmp_name, seg_name)
 
@@ -70,6 +126,20 @@ class Segmenter:
 
         corpus = "\n".join(corpus)
         with open(self.seg_join_dir+"/corpus.txt", "w") as seg_join_file:
+            seg_join_file.write(corpus)
+
+    # 将分词后的每个小文件代表一个文档，将这些文件拼接成一个大文件，大文件每行代表一个文档
+    def join_pos_seg_file(self):
+        corpus = []
+        for i in range(1, self.doc_count + 1):
+            seg_name = "%s/pos_seg/%d" % (self.proj_name, i)
+            with open(seg_name, "r") as seg_file:
+                lines = [line.strip() for line in seg_file.readlines() if len(line.strip()) > 0]
+                doc = " ".join(lines)
+                corpus.append(doc)
+
+        corpus = "\n".join(corpus)
+        with open(self.seg_join_dir + "/corpus1.txt", "w") as seg_join_file:
             seg_join_file.write(corpus)
 
     # 分句
@@ -92,6 +162,7 @@ class Segmenter:
         return sentences
 
 if __name__ == "__main__":
-    segmenter = Segmenter("article150801160830")
-    segmenter.seg()
-    segmenter.join_seg_file()
+    segmenter = Segmenter("article_cat")
+    # segmenter.seg_title_tags()
+    segmenter.pos_seg()
+    segmenter.join_pos_seg_file()
